Push logic down into TrustyChart
